Added initial version of client-server update engine.
Augmented the BrowserInstance class with toJSON(), genInitInstructions(),
and helper functions.  toJSON() uses the other methods to serialize the DOM
for a BrowserInstance into a JSON object containing instructions for the client
side update engine.  Added client.js, which is the start of the client side
library.  client.js houses a vt global object/namespace with a start() method
and an UpdateEngine class.  The UpdateEngine transforms JSON instructions
into DOM updates.  Added a test in test-browser-instance that does a very
basic test of the protocol.  All tests pass.