Added .nodeStats() to domutils, plus a test.
nodeStats() is for debug purposes.  It prints out a tally of the types
of nodes in a BrowserInstance.  test-nodestats.js connects to 4 web sites
and prints out their nodeStats().  I did this to get a sense of which w3c
node types are actually used in HTML.

Note: also added test/inactive.  Tests in the inactive folder don't get
run by run_tests.js.