Added vt.Server class, which is an API for creating HTTP servers that uses BrowserInstances.
vt.Server is in lib/server.js.  It is the highest level API for our
library, and it manages a Connect server that uses BrowserInstances
to load/render pages.  The user just needs to pass in a router function
and some configuration parameters.  This will also be useful for
programmatically creating a server for unit tests.

Rewrote examples/test-server/server.js to use the new Server class.