Finished initial support for multiple clients per BrowserInstance.
Added /sharedbrowsing route to examples server for testing.

Changed BrowserInstance from inheritance.js to a regular JavaScript class
so we can use Node's sys.inherit to make BrowserInstance an EventEmitter.

BrowserInstance now emits a 'load' event, which we make use of to queue
socket.io connections until the BrowserInstance is loaded.

Instead of always using the request's sessionID to identify the browser,
the ID can be configured by the client programmer.  This is required for
2 people to share the same browser, since they won't have identical session
IDs.