Big update.  Now using aspect oriented approach for detecting DOM updates.
Deleted jsdom-adapter.js and moved JSDOM specific code to dom.js.

We now thinly wrap the DOM API on the server, and then execute the same
functions on the client. In the near future, we will use jQuery to implement
the client side, for cross browser compatability, but for now, we just invoke
the DOM function.

dom.js has the DOM class, which manages the JSDOM DOM and emits EventEmitter
events whenever the DOM is manipulated (detected via interposing on DOM
methods).  A JSON-RPC compliant message is passed to event listeners.
Sending the message to the client will make the client perform the same
modification on their DOM.

Added methods to update_engine.js to add the DOM API functions to our JSON-RPC
protocol.

Updated BrowserInstance to listen for 'DOMManipulation' events from dom.js, and
then rethrow them.

Updated ClientManager to listen for 'DOMManipulation' events from the
BrowserInstance.  The event listener receives the JSON-RPC object, stringifies
it, and broadcasts it to all connected clients.

We are still using the old protocol for the initial sync().  Converting that
is next.