First steps toward implementing asynchronous peer notification protocol.
Added NotifyPeer, which is currently implemented on the server.  A NotifyPeer
instance makes all of the functions in the "api" object passed in available
over JSON-RPC notifications.  JSON-RPC notifications are RPC calls that do not
return anything.  This means we don't have to keep track of asynchronous return
values.  NotifyPeer should work over any socket connection, and we can use the
same code in the browser (to implement the update engine API) and the server
(to implement the client side event processing API).