Now serving client code via browserify, for browser-side require.
This makes it much easier to organize code (require()s instead of
closures), and also much easier to share code between the client and
server (e.g. NotifyPeer).  Browserify plugs straight into our Connect
middleware, and has support for minifying.

This is the first of many steps in getting the client side code
organized and plugged into the test framework, which is part of an
over-arching project to make the client/server communication framework
more well-defined and robust.

This update requires browserify@0.3.1: npm install browserify@0.3.1