Zombies can now follow links and sync the BrowserInstances.
We now use zombie's loaded event, and the loading event we added to
zombie.  While loading, we queue up events, and then when loaded, we
dispatch those events to the clients.  We are missing some functions in
our advice, because some text nodes aren't getting set properly when we
change pages.