Added XMLHttpRequest support and fixed some bugs.
XMLHttpRequest lives in server/XMLHttpRequest.js and is attached to the
window in server/browser.coffee.

Now attaching console to the window in browser.coffee.

Fixed some bugs in jsdomwrapper.coffee.

window.location is now being set properly in browser.coffee.  Not
setting this was causing jQuery's AJAX load function to fail.

Switched the client back to only sending clicks.  Sending mousemove and
friends was causing tons of debug output.  We should make this
configurable (or detect which events the server DOM code actually waits
on), because sending things like mousemove sends a lot of packets.