Added jsdom as a git submodule in the deps/ folder.
The jsdom version in npm is 0.2.0, which was released in February.  To
take advantage of all the fixes since then, we're now tracking the jsdom
git repository directly.  You should symlink deps/jsdom into
node_modules/jsdom.  The Cakefile will be updated to do this
automatically when running "cake setup".