Added a partial w3c spec Location implementation.
window.location now interacts with an actual location object.  This
object deals with resolving URLs, triggering hashchange when required,
and calling browser.load when appropriate.

Cleaned up and refactored Browser, whose load method is now much
shorter.

The idea is that Browser#load is like typing a URL in a menubar and
hitting enter.  Using window.location should be just like using it in a
normal browser.  When window.location causes the browser to navigate to
another page, it uses Browser#load to do it.

Moved createWindow() from Browser to JSDOMWrapper, and cut out the hacky
Location implementaiton in favor of the new Location object.