Fix for advice when an iframe calls setAttribute.
In the upcoming frame improvements for JSDOM, frames
Documents/DOMWindows aren't created until the frame is inserted into the
DOM (as per HTML5 spec).  This means we have to wait for the iframe
'load' event before we can tag the new Document, since it won't be
created immediately after calling setAttribute like it was before the
updates.