Migrated from DNode to Socket.io.
DNode relies on an older version of socket.io (0.6.x).  Chrome 14 and
FireFox6 have updated to the newest WebSocket spec, which is drastically
different than the old spec.  Older versions of Socket.io don't support
the new WebSockets, but the newer versions do (0.8.x).

The changes between Socket.IO versions are non-trivial, so updating
DNode will be involved.  Socket.IO >= 0.7.x also includes features that
overlap with DNode, so it made sense to migrate to just using Socket.IO,
so we can stay up to date with changing protocols easier.

This commit is mostly a trancription from DNode to Socket.io.  There are
probably places we can refactor to take advantage of the new Socket.io
features.