Finished form support.  Added 'setProperty' on client.
Some properties need to be intercepted, in addition to attributes.
'selected' is the only current one used.  If we don't set it as a
property on the client, nothing happens.  Also, setting it as a property
on the server doesn't call setNamedItem, since it doesn't actually set
an attribute (although attributes set with setAttribute do set
properties).