Added bin/jquery-test to run jquery test suite using phantom.
Requires phantomjs to be in PATH.

Requires deps/jquery-cloudbrowser to be available at
localhost/jquery-cloudbrowser, OR, you can specify the URL for it as a
parameter to jquery-test.  You need to use this version, which has
QUnit hooks for reporting test info on the command line.

You can specify which test suite to run:
    bin/jquery-test traversing

It runs all tests by default.  If you need to pass a custom URL and want
to run all tests, pass all:
    bin/jquery-test all http://localhost/some/other/path/test/