Fix for socket.io-client version mismatch.
node_modules/socket.io depends on socket.io-client, and since we were
installing deps/socket.io-client to node_modules, it would use that
instead of installing its own.  This caused weird
disconnection/reconnection issues and client-side errors.

To fix it, we don't install deps/socket.io-client to node_modules, and
just require it directly when using it for our tests (which is all it's
there for).  We use the git version because the latest npm version
doesn't work.