Removed the need to require multiple JSDOM instances.
This required changes to JSDOM and CloudBrowser

jsdom.jsdom now takes a 'browser' argument, which is a CloudBrowser
Browser object.  This gets passed to new Documents, and our advice that
wraps the Document constructor embeds a reference to the browser in the
Document.  The advice is added only once, and each method looks up the
browser on the current Document object (chasing references such as
node._ownerDocument).

This drastically reduces memory usage for virtual browsers.